+++ src/main/java/com/kobe/warehouse/service/dto/builder/VenteStatQueryBuilder.java
@@ -24,7 +24,6 @@ public final class VenteStatQueryBuilder {
   public static final String PERIOQIQUE_CA_QUERY =
       "SELECT SUM(s.sales_amount) AS sales_amount,SUM(s.ht_amount)/COUNT(s.id) AS panierMoyen,SUM(s.amount_to_be_paid) AS amount_to_be_paid,SUM(s.discount_amount) AS discount_amount,SUM(s.cost_amount) AS cost_amount, SUM(s.ht_amount)-SUM(s.cost_amount) AS marge, SUM(s.amount_to_be_taken_into_account) AS amount_to_be_taken_into_account,SUM(s.net_amount) AS net_amount,SUM(s.ht_amount) AS ht_amount,SUM(s.part_assure) AS part_assure,SUM(s.part_tiers_payant) AS part_tiers_payant,SUM(s.tax_amount) AS tax_amount,SUM(s.rest_to_pay) AS rest_to_pay,SUM(s.ht_amount_ug) AS ht_amount_ug,SUM(s.discount_amount_hors_ug) AS discount_amount_hors_ug,SUM(s.discount_amount_ug) AS discount_amount_ug,SUM(s.net_ug_amount) AS net_ug_amount,SUM(s.marge_ug) AS marge_ug,SUM(s.montant_ttc_ug) AS montant_ttc_ug,SUM(s.payroll_amount) AS payroll_amount,SUM(s.montant_tva_ug) AS montant_tva_ug,SUM(s.montant_net_ug) AS montant_net_ug,SUM(payment.paid_amount) AS paid_amount,SUM(payment.real_net_amount) AS real_net_amount,COUNT(s.id) AS sale_count FROM sales s,(SELECT SUM(p.net_amount) AS real_net_amount,SUM(p.paid_amount) AS paid_amount,p.sales_id AS sales_id FROM payment p GROUP BY p.sales_id) as payment WHERE payment.sales_id=s.id   AND s.imported=0 AND  DATE(s.updated_at) BETWEEN ?1 AND ?2 AND s.statut IN (?3)  %s %s %s";
 
-
   public static final String TYPE_VENTE = " AND s.dtype ='%s' ";
   public static final String PERIOQIQUE_CA_QUERY_GROUPING_BY_PERIODE =
       "SELECT %s AS mvtDate,SUM(s.ht_amount)/COUNT(s.id) AS panierMoyen, s.statut AS statut, SUM(s.sales_amount) AS sales_amount,SUM(s.amount_to_be_paid) AS amount_to_be_paid,SUM(s.discount_amount) AS discount_amount,SUM(s.cost_amount) AS cost_amount, SUM(s.ht_amount)-SUM(s.cost_amount) AS marge, SUM(s.amount_to_be_taken_into_account) AS amount_to_be_taken_into_account,SUM(s.net_amount) AS net_amount,SUM(s.ht_amount) AS ht_amount,SUM(s.part_assure) AS part_assure,SUM(s.part_tiers_payant) AS part_tiers_payant,SUM(s.tax_amount) AS tax_amount,SUM(s.rest_to_pay) AS rest_to_pay,SUM(s.ht_amount_ug) AS ht_amount_ug,SUM(s.discount_amount_hors_ug) AS discount_amount_hors_ug,SUM(s.discount_amount_ug) AS discount_amount_ug,SUM(s.net_ug_amount) AS net_ug_amount,SUM(s.marge_ug) AS marge_ug,SUM(s.montant_ttc_ug) AS montant_ttc_ug,SUM(s.payroll_amount) AS payroll_amount,SUM(s.montant_tva_ug) AS montant_tva_ug,SUM(s.montant_net_ug) AS montant_net_ug,SUM(payment.paid_amount) AS paid_amount,SUM(payment.real_net_amount) AS real_net_amount,COUNT(s.id) AS sale_count FROM sales s,(SELECT SUM(p.net_amount) AS real_net_amount,SUM(p.paid_amount) AS paid_amount,p.sales_id AS sales_id FROM payment p GROUP BY p.sales_id) as payment WHERE payment.sales_id=s.id   AND s.imported=0 AND  DATE(s.updated_at) BETWEEN ?1 AND ?2 AND s.statut IN (?3) %s %s %s GROUP BY mvtDate,statut ORDER BY mvtDate ";
@@ -65,9 +64,11 @@ FROM payment p,sales s,payment_mode md WHERE p.payment_mode_code=md.code AND p.s
         tuple.get("montant_net_ug", BigDecimal.class),
         tuple.get("paid_amount", BigDecimal.class),
         tuple.get("real_net_amount", BigDecimal.class),
-        tuple.get("sale_count", BigInteger.class),
-        tuple.get("panierMoyen", BigDecimal.class).round(new MathContext(2, RoundingMode.HALF_UP)).doubleValue()
-        );
+        tuple.get("sale_count", Long.class),
+        tuple
+            .get("panierMoyen", BigDecimal.class)
+            .round(new MathContext(2, RoundingMode.HALF_UP))
+            .doubleValue());
   }
 
   public static VenteByTypeRecord buildVenteByTypeRecord(Tuple tuple) {
